requirements-simplified.txt
fastapi>=0.104.1
pydantic>=2.4.2
uvicorn>=0.23.2
python-dotenv>=1.0.0
python-multipart>=0.0.6